# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enterprise OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - in: query
          name: employee_id
          schema:
            type: integer
        - in: query
          name: date
          schema:
            type: string
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: createAttendance
      tags: [attendance]
      summary: Create attendance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"

  # --- LEAVES ---
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update contact
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete contact
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- DEALS ---
  /deals:
    get:
      operationId: listDeals
      tags: [deals]
      summary: List deals
      parameters:
        - in: query
          name: stage
          schema:
            type: string
      responses:
        "200":
          description: Deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
    post:
      operationId: createDeal
      tags: [deals]
      summary: Create deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}:
    patch:
      operationId: updateDeal
      tags: [deals]
      summary: Update deal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
    delete:
      operationId: deleteDeal
      tags: [deals]
      summary: Delete deal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- PRODUCTS (ERP) ---
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Send message to AI copilot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopilotResponse"
  /ai/suggestions:
    get:
      operationId: getAiSuggestions
      tags: [ai]
      summary: Get contextual AI prompt suggestions
      parameters:
        - in: query
          name: module
          schema:
            type: string
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiSuggestion"

  # --- WORKFLOWS ---
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
    AuthResponse:
      type: object
      required: [access_token, token_type, user]
      properties:
        access_token:
          type: string
        token_type:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatar:
          type: ["string", "null"]
        created_at:
          type: string
...
    AttendanceRecord:
      type: object
      required: [id, employee_id, employee_name, date, check_in, status]
      properties:
        id:
          type: integer
        employee_id:
          type: integer
        employee_name:
          type: string
        date:
          type: string
        check_in:
          type: ["string", "null"]
        check_out:
          type: ["string", "null"]
        status:
          type: string
    AttendanceInput:
      type: object
      required: [employee_id, date, status]
      properties:
        employee_id:
          type: integer
        date:
          type: string
        check_in:
          type: string
        check_out:
          type: string
        status:
          type: string

    # --- LEAVES ---
...
    Contact:
      type: object
      required: [id, name, email, company, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        company:
          type: string
        role:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        created_at:
          type: string
    ContactInput:
      type: object
      required: [name, email, company]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        role:
          type: string
    ContactUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        role:
          type: string

    # --- DEALS ---
    Deal:
      type: object
      required: [id, title, contact, value, stage, close_date, created_at]
      properties:
        id:
          type: integer
        title:
          type: string
        contact:
          type: string
        company:
          type: ["string", "null"]
        value:
          type: number
        stage:
          type: string
        probability:
          type: ["number", "null"]
        close_date:
          type: string
        assigned_to:
          type: ["string", "null"]
        created_at:
          type: string
    DealInput:
      type: object
      required: [title, contact, value, stage, close_date]
      properties:
        title:
          type: string
        contact:
          type: string
        company:
          type: string
        value:
          type: number
        stage:
          type: string
        probability:
          type: number
        close_date:
          type: string
        assigned_to:
          type: string
    DealUpdate:
      type: object
      properties:
        title:
          type: string
        contact:
          type: string
        company:
          type: string
        value:
          type: number
        stage:
          type: string
        probability:
          type: number
        close_date:
          type: string
        assigned_to:
          type: string

    # --- PRODUCTS (ERP) ---
...
    AiChatInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        module:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
        content:
          type: string
    CopilotResponse:
      type: object
      required: [response, module]
      properties:
        response:
          type: string
        module:
          type: string
        suggestions:
          type: array
          items:
            type: string
    AiSuggestion:
      type: object
      required: [id, prompt, category]
      properties:
        id:
          type: integer
        prompt:
          type: string
        category:
          type: string

    # --- WORKFLOWS ---
...